# docs/swagger.yaml
...
openapi: 3.0.0
info:
  title: SSU SOCIAL API
  description: MongoDB ExpressJS NodeJS REST API
  version: 1.0.0
servers:
  - url: http://localhost:8095/
    description: Localhost development server 
  - url: https://ssu-social-app-api.onrender.com/
    description: Render remote server
components:
  securitySchemes:
    ApiKeyAuth:        # arbitrary name for the security scheme
      type: apiKey
      in: header       # can be "header", "query" or "cookie"
      name: auth-token
  schemas:
...
    Post:
      type: object
      properties:
        username:
          type: string
        content:
          type: string
...
paths:
...
  /feed/{startingPosition}/{pageSize}:                
    get:
      summary: Retrieves a portion of a list of posts that is appropriate for an anonymous user.
      description: Retrieves a list of posts as JSON objects.  The posts are sorted to present posts with higher likes and views first.
      tags:
        - Feed
      parameters:
        - in: path
          name: startingPosition
          required: true
          description: zero-based index of the first post to return.
          schema:
            type: integer
        - in: path
          name: pageSize
          required: true
          description: total number of posts returned.
          schema:
            type: integer
      responses:
        200:
          description: A list of posts JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
  /feed/{username}:                
    get:
      summary: Retrieves a list of posts that is appropriate for a given user.
      description: Retrieves a list of posts as JSON objects.  The posts are sorted to present posts with higher likes and views first.
      tags:
        - Feed
      parameters:
        - in: path
          name: username
          required: true
          description: username.
          schema:
            type: string
      responses:
        200:
          description: A list of posts JSON objects in an array.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Post"
#
# Project Notes
#
...